requirements.txt
gradio
requests
python-dotenv
Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope CurrentUser